bot/requirements.txt
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
discord.py==1.4.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
idna-ssl==1.1.0
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
multidict==4.7.6
netifaces==0.10.4
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.5.3
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyserial==3.4
python-debian==0.1.32
python-dotenv==0.13.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
ssh-import-id==5.7
Twisted==17.9.0
urllib3==1.22
websockets==8.1
yarl==1.5.1
